Centralize Qt initialization
